Prototype statistical info and visualization per region
The geostats series of file presents some ideas on how to sumarize professional
info per distinct administrative level. For now, each region is treated in an
isolated fashion. So, we are unable to depict border effects. This should bt
addressed ASAP. Aims are:

 - Add corrections and counts per region population/size;
 - Add useful annotations (e. g. most important city);
 - Put brazilian mesh to work;

We found that some shapefiles contain errors. A way of curating those files is
urgently needed, too. A tempory solution made with QGIS is available.